Adds Ecto-backed queue functionality configurable in application.ex
